style: devide clone.py to create_training_data.py & train_model.py
